requirements.txt
numpy==1.18.5
opencv-python==4.4.0.46
torch==1.6.0
torchvision==0.7.0
Pillow==8.0.1
visdom==0.1.8.9
dominate==2.6.0
yacs==0.1.8
scikit-image==0.17.2
face-alignment
chumpy==0.70